requirements.txt
opencv-python
numpy
torch
torchvision
pillow
transformers